+++ apps/flutter_architecture/lib/app/router/app_router.dart
@@ -1,6 +1,7 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:flutter_architecture/app/router/auth_guard.dart';
 import 'package:flutter_architecture/features/auth/presentation/pages/login_page.dart';
+import 'package:flutter_architecture/features/catalog/presentation/pages/catalog_page.dart';
 import 'package:flutter_architecture/features/profile/presentation/pages/profile_page.dart';
 import 'package:flutter_architecture/features/protected/presentation/pages/protected_page.dart';
 import 'package:flutter_architecture/features/shell/presentation/pages/shell_page.dart';
@@ -19,8 +20,9 @@ part 'app_router.gr.dart';
 ///   ↓
 /// ShellRoute(A)
 ///   ├── LoginRoute(B)
-///   ├── ProfileRoute(C)
-///   └── ProtectedRoute(D, guarded)
+///   ├── CatalogRoute(C)
+///   ├── ProfileRoute(D)
+///   └── ProtectedRoute(E, guarded)
 /// ```
 @lazySingleton
 @AutoRouterConfig(replaceInRouteName: 'Page,Route')
@@ -31,17 +33,15 @@ class AppRouter extends RootStackRouter {
 
   @override
   List<AutoRoute> get routes => <AutoRoute>[
-        AutoRoute(
-          page: ShellRoute.page,
-          initial: true,
-          children: <AutoRoute>[
-            AutoRoute(page: LoginRoute.page, initial: true),
-            AutoRoute(page: ProfileRoute.page),
-          ],
-        ),
-        AutoRoute(
-          page: ProtectedRoute.page,
-          guards: <AutoRouteGuard>[_authGuard],
-        ),
-      ];
+    AutoRoute(
+      page: ShellRoute.page,
+      initial: true,
+      children: <AutoRoute>[
+        AutoRoute(page: LoginRoute.page, initial: true),
+        AutoRoute(page: CatalogRoute.page),
+        AutoRoute(page: ProfileRoute.page),
+      ],
+    ),
+    AutoRoute(page: ProtectedRoute.page, guards: <AutoRouteGuard>[_authGuard]),
+  ];
 }